Fix and re-format
