+++ app/src/main/java/Objects/ResultStatus.java
@@ -1,4 +1,4 @@
-package TransferObjects;
+package Objects;
 
 public enum ResultStatus
 {
