+++ app/src/main/java/Objects/genericTask.java
@@ -1,0 +1,13 @@
+package Objects;
+
+/*
+    Generic type for tasks in order to standardise across different API calls
+    All info received from API calls should be entered into this class
+*/
+public class genericTask
+{
+    private String title;
+    private String description;
+    private double longitude;
+    private double latitude;
+}
