+++ app/src/main/java/apps/mobile/ostium/GPSModule.java
@@ -4,8 +4,8 @@ import android.annotation.SuppressLint;
 import android.location.Location;
 import android.location.LocationListener;
 import android.location.LocationManager;
-import TransferObjects.GetLocationRequest;
-import TransferObjects.ResultStatus;
+import Objects.GetLocationRequest;
+import Objects.ResultStatus;
 
 public class GPSModule
 {
