+++ app/src/main/java/apps/mobile/ostium/androidCalendarHandle.java
@@ -9,15 +9,8 @@ import android.os.Bundle;
 import android.os.ResultReceiver;
 import android.provider.CalendarContract;
 
-
-public class androidCalendarHandle extends IntentService {
-
-
-    /**
-     * Creates an IntentService.  Invoked by your subclass's constructor.
-     *
-     * @param name Used to name the worker thread, important only for debugging.
-     */
+public class androidCalendarHandle extends IntentService
+{
     public androidCalendarHandle(String name) {
         super(name);
     }
@@ -26,14 +19,14 @@ public class androidCalendarHandle extends IntentService {
         super("androidCalendarHandle");
     }
 
-    //Invoked on Intent Service Start
+    // Invoked on Intent Service Start
     @Override
-    protected void onHandleIntent(Intent intent) {
-
-        //Get ResultReceiver from the Intent passed to service
+    protected void onHandleIntent(Intent intent)
+    {
+        // Get ResultReceiver from the Intent passed to service
         final ResultReceiver receiver = intent.getParcelableExtra("receiver");
 
-        //Create bundle to store data to send back
+        // Create bundle to store data to send back
         Bundle bundle = new Bundle();
 
         /*
@@ -41,22 +34,21 @@ public class androidCalendarHandle extends IntentService {
         * TODO: Implement success and fail cases for receiver.send
         * */
         receiver.send(1, bundle);
-
     }
 
-    public void getDataFromEventTable() {
-
-        Cursor cur = null;
+    public void getDataFromEventTable()
+    {
+        Cursor cur;
         ContentResolver cr = getContentResolver();
 
         String[] mProjection =
-                {
-                        "_id",
-                        CalendarContract.Events.TITLE,
-                        CalendarContract.Events.EVENT_LOCATION,
-                        CalendarContract.Events.DTSTART,
-                        CalendarContract.Events.DTEND,
-                };
+            {
+                "_id",
+                CalendarContract.Events.TITLE,
+                CalendarContract.Events.EVENT_LOCATION,
+                CalendarContract.Events.DTSTART,
+                CalendarContract.Events.DTEND,
+            };
 
         Uri uri = CalendarContract.Events.CONTENT_URI;
         String selection = CalendarContract.Events.EVENT_LOCATION + " = ? ";
@@ -64,10 +56,10 @@ public class androidCalendarHandle extends IntentService {
 
         cur = cr.query(uri, mProjection, selection, selectionArgs, null);
 
-        while (cur.moveToNext()) {
-
-            //EXAMPLE: Get title
-            String title = cur.getString(cur.getColumnIndex(CalendarContract.Events.TITLE));
+        while (cur.moveToNext())
+        {
+            // EXAMPLE: Get title
+            // String title = cur.getString(cur.getColumnIndex(CalendarContract.Events.TITLE));
 
             //TODO: Handle data
         }
