--- app/src/main/java/apps/mobile/ostium/genericTask.java
@@ -1,13 +1,0 @@
-package apps.mobile.ostium;
-
-/**
- * Generic type for tasks in order to standardise across different API calls
- *
- * All info received from API calls should be entered into this class
- */
-public class genericTask {
-    private String title;
-    private String description;
-    private double longitude;
-    private double latitude;
-}
